+++ src/ui/button/button-primary/index.ts
@@ -1,0 +1,1 @@
+export * from './button-primary.component'
