+++ src/ui/button/button-search/button-search.component.tsx
@@ -1,0 +1,9 @@
+import { FC } from 'react'
+import { ButtonProps } from '../button.interfaces'
+import styles from './button-search.module.css'
+
+export const ButtonSearch: FC<ButtonProps> = ({ text, onClick }) => (
+  <button className={styles.buttonSearch} onClick={onClick}>
+    <span className={styles.buttonTextSearch}>{text}</span>
+  </button>
+)
