+++ src/ui/button/button-search/index.ts
@@ -1,0 +1,1 @@
+export * from './button-search.component'
