+++ src/ui/button/go-back-button/index.ts
@@ -1,0 +1,1 @@
+export * from './go-back-button.component'
