+++ src/ui/button/index.ts
@@ -1,1 +1,3 @@
-export * from './button.component'
+export * from './button-search'
+export * from './button-primary'
+export * from './go-back-button'
